lanhackbackend/requirements.txt
fastapi
paho-mqtt